enterprise_sdk/pyproject.toml
[project]
name = "encypher-enterprise"
version = "1.0.0"
description = "Python SDK for Encypher Enterprise API - C2PA content signing with streaming support"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "Proprietary"}
authors = [
    {name = "EncypherAI", email = "sdk@encypherai.com"}
]
keywords = ["c2pa", "content-signing", "provenance", "ai", "streaming", "langchain", "openai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "gitpython>=3.1.45",
    "pyyaml>=6.0.3",
    "langdetect>=1.0.9",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "pillow>=11.3.0",
]

[project.optional-dependencies]
langchain = ["langchain>=0.1.0", "langchain-core>=0.1.0"]
openai = ["openai>=1.0.0"]
litellm = ["litellm>=1.0.0"]
all = ["langchain>=0.1.0", "langchain-core>=0.1.0", "openai>=1.0.0", "litellm>=1.0.0"]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://encypherai.com"
Documentation = "https://docs.encypherai.com/sdk"
Repository = "https://github.com/encypherai/encypherai-commercial"

[project.scripts]
encypher = "encypher_enterprise.cli.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "--cov=encypher_enterprise --cov-report=html --cov-report=term"

[tool.black]
line-length = 100
target-version = ['py39']

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
